Fixing bugs

+++ To-Do-Management_front/todo-frontend/src/pages/Dashboard.tsx
@@ -1,19 +1,21 @@
-// src/pages/Dashboard.tsx
-import {useEffect, useRef, useState} from "react";
+
+import {useEffect, useState} from "react";
 import type { CreateTaskRequest, TodoTask, UpdateTaskRequest } from "../types";
 import { createTask, deleteTask, getTasks, updateTask } from "../api/todoApi";
 import { TaskForm } from "../components/TaskForm";
 import { TasksTable } from "../components/TaskCard";
-import { Toast } from 'primereact/toast';
+import {ConfirmDialog, confirmDialog} from "primereact/confirmdialog";
+import {useToast} from "../context/ToastContext.tsx";
 
 const Dashboard = () => {
+    const {show} = useToast();
+
     const [tasks, setTasks] = useState<TodoTask[]>([]);
     const [loading, setLoading] = useState(true);
     const [editingTask, setEditingTask] = useState<TodoTask | null>(null);
     const [page, setPage] = useState(1);
-    const [totalPages, setTotalPages] = useState(1);
-    const pageSize = 12;
-    const toast = useRef<Toast>(null);
+    const [totalCount, setTotalCount] = useState(0);
+    const pageSize = 10;
 
 
     const fetchTasks = async (pageNumber: number) => {
@@ -21,7 +23,7 @@ const Dashboard = () => {
         try {
             const res = await getTasks(pageNumber, pageSize);
             setTasks(res.data);
-            setTotalPages(Math.ceil(res.totalCount / pageSize));
+            setTotalCount(res.totalCount);
         } catch (err) {
             console.error(err);
         } finally {
@@ -34,14 +36,25 @@ const Dashboard = () => {
     }, [page]);
 
     const handleDelete = async (id: number) => {
-        try {
-            await deleteTask(id);
-            fetchTasks(page);
-            toast.current?.show({ severity: 'success', summary: 'Deleted', detail: 'Task deleted successfully', life: 2000 });
-        } catch (err) {
-            console.error(err);
-            toast.current?.show({ severity: 'error', summary: 'Error', detail: 'Failed to delete task.', life: 3000 });
-        }
+        confirmDialog({
+            message: 'Do you want to delete this task?',
+            header: 'Delete Confirmation',
+            icon: 'pi pi-exclamation-triangle',
+            acceptClassName: 'p-button-danger',
+            accept: async () => {
+                try {
+                    await deleteTask(id);
+                    fetchTasks(page);
+                    show({ severity: 'success', summary: 'Deleted', detail: 'Task deleted successfully', life: 3000 });
+                } catch (err) {
+                    console.error(err);
+                    show({ severity: 'error', summary: 'Error', detail: 'Failed to delete task.', life: 3000 });
+                }
+            },
+            reject: () => {
+                show({ severity: 'info', summary: 'Cancelled', detail: 'Delete cancelled', life: 3000 });
+            }
+        });
     };
 
     const handleSave = async (taskData: CreateTaskRequest | UpdateTaskRequest) => {
@@ -50,32 +63,30 @@ const Dashboard = () => {
                 const updated = await updateTask(editingTask.id, taskData as UpdateTaskRequest);
                 setTasks(prev => prev.map(t => t.id === updated.id ? updated : t));
                 setEditingTask(null);
-                toast.current?.show({ severity: 'success', summary: 'Updated', detail: 'Task updated successfully', life: 3000 });
+                show({ severity: 'success', summary: 'Updated', detail: 'Task updated successfully', life: 3000 });
             } else {
                 await createTask(taskData as CreateTaskRequest);
                 fetchTasks(page);
-                toast.current?.show({ severity: 'success', summary: 'Created', detail: 'Task created successfully', life: 3000 });
+                show({ severity: 'success', summary: 'Created', detail: 'Task created successfully', life: 3000 });
             }
         } catch (err: any) {
             if (err.response?.status === 409) {
-                toast.current?.show({ severity: 'error', summary: 'Error', detail: 'Task with this title already exists!', life: 3000 });
+                show({ severity: 'error', summary: 'Error', detail: 'Task with this title already exists!', life: 3000 });
             } else if (err.response?.status === 400 && Array.isArray(err.response.data)) {
                 const messages = err.response.data.map(
                     (e: { propertyName: string; errorMessage: string }) =>
                         `${e.propertyName}: ${e.errorMessage}`
                 ).join("\n");
-                toast.current?.show({ severity: 'error', summary: 'Validation Error', detail: messages, life: 3000 });
+                show({ severity: 'error', summary: 'Validation Error', detail: messages, life: 3000 });
             } else {
-                toast.current?.show({ severity: 'error', summary: 'Error', detail: 'Something went wrong.', life: 3000 });
+                show({ severity: 'error', summary: 'Error', detail: 'Something went wrong.', life: 3000 });
             }
         }
     };
 
-    if (loading) return <p>Loading...</p>;
-
     return (
         <div style={{ padding: "20px" }}>
-            <Toast ref={toast} />
+            <ConfirmDialog />
 
             <h1 style={{ textAlign: "center" }}>Dashboard</h1>
 
@@ -89,6 +100,11 @@ const Dashboard = () => {
                 tasks={tasks}
                 onEdit={(task) => setEditingTask(task)}
                 onDelete={(task) => handleDelete(task.id)}
+                paginator
+                rows={pageSize}
+                totalRecords={totalCount}
+                first={(page - 1) * pageSize}
+                onPage={(e) => setPage(e.first / e.rows + 1)}
             />
         </div>
     );
